+++ docs/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Added
+
+-   A method to run a subprocess, and show the real time stdout and stderr and capture it
+
 ## [0.4.0] - 2024-05-17
 
 ### Added